refactor: move auto trigger coefficients to separate json file.
This requires an explicit include of the JSON file in the tsconfig.
